requirements.txt
# =============================================================================
# CORE DJANGO
# =============================================================================
Django==5.2.8
python-decouple

# =============================================================================
# DJANGO REST FRAMEWORK
# =============================================================================
djangorestframework
django-filter

# =============================================================================
# AUTHENTICATION & AUTHORIZATION
# =============================================================================
djangorestframework-simplejwt
django-allauth
dj-rest-auth

# =============================================================================
# CORS
# =============================================================================
django-cors-headers

# =============================================================================
# DJANGO CHANNELS (WebSocket)
# =============================================================================
channels
channels-redis
daphne

# =============================================================================
# CELERY & REDIS
# =============================================================================
celery
redis
django-redis
django-celery-beat
kombu
vine

# =============================================================================
# DATABASE
# =============================================================================
psycopg2-binary

# =============================================================================
# IMAGE PROCESSING
# =============================================================================
Pillow

# =============================================================================
# API DOCUMENTATION
# =============================================================================
drf-spectacular[sidecar]

# =============================================================================
# STATIC FILES
# =============================================================================
whitenoise

# =============================================================================
# PRODUCTION SERVER
# =============================================================================
gunicorn

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv
pytz
requests

# =============================================================================
# DEVELOPMENT TOOLS
# =============================================================================
django-extensions
django-debug-toolbar

# =============================================================================
# TESTING
# =============================================================================
pytest
pytest-django
pytest-cov
factory-boy
faker

# =============================================================================
# CODE QUALITY
# =============================================================================
black
flake8
isort
pylint
pylint-django

# =============================================================================
# MONITORING & ERROR TRACKING (Optional - Production)
# =============================================================================
sentry-sdk

# =============================================================================
# CLOUD STORAGE (Optional - Production)
# =============================================================================
# boto3
# django-storages

# =============================================================================
# SECURITY (Optional - Production)
# =============================================================================
# django-ratelimit
# django-axes
